+++ videocaptioner/ui/components/transcription_setting_card.py
@@ -12,6 +12,7 @@ from videocaptioner.core.entities import (
 from videocaptioner.core.utils.platform_utils import is_macos
 
 from .FasterWhisperSettingWidget import FasterWhisperSettingWidget
+from .MiniMaxAPISettingWidget import MiniMaxAPISettingWidget
 from .WhisperAPISettingWidget import WhisperAPISettingWidget
 from .WhisperCppSettingWidget import WhisperCppSettingWidget
 
@@ -32,6 +33,7 @@ class TranscriptionSettingCard(QWidget):
         self.empty_widget = QWidget(self)  # 添加空白页面作为默认显示
         self.whisper_cpp_widget = WhisperCppSettingWidget(self)
         self.whisper_api_widget = WhisperAPISettingWidget(self)
+        self.minimax_api_widget = MiniMaxAPISettingWidget(self)
 
         # FasterWhisper 在 macOS 上不可用
         self.faster_whisper_widget: Optional[FasterWhisperSettingWidget] = None
@@ -41,6 +43,7 @@ class TranscriptionSettingCard(QWidget):
         self.stacked_widget.addWidget(self.empty_widget)  # 添加空白页面
         self.stacked_widget.addWidget(self.whisper_cpp_widget)
         self.stacked_widget.addWidget(self.whisper_api_widget)
+        self.stacked_widget.addWidget(self.minimax_api_widget)
         if self.faster_whisper_widget is not None:
             self.stacked_widget.addWidget(self.faster_whisper_widget)
 
@@ -52,6 +55,8 @@ class TranscriptionSettingCard(QWidget):
             self.stacked_widget.setCurrentWidget(self.whisper_cpp_widget)
         elif value == TranscribeModelEnum.WHISPER_API.value:
             self.stacked_widget.setCurrentWidget(self.whisper_api_widget)
+        elif value == TranscribeModelEnum.MINIMAX_API.value:
+            self.stacked_widget.setCurrentWidget(self.minimax_api_widget)
         elif value == TranscribeModelEnum.FASTER_WHISPER.value:
             self.stacked_widget.setCurrentWidget(self.faster_whisper_widget)
         else: